Show only full description if available
The "abstract" and the "description" are two very similar teaser about
the movie or series. Sometimes it's annoying to read the "abstract" then
read (almost) the same text when the description begins.

Instead of showing both the "abstract" and the "description", show only
"description" if available, and show only the "abstract" if no
description found.